refactor<mutations>: add index files for mutations directory
