+++ src/graphql/mutations/blog/index.js
@@ -1,0 +1,1 @@
+export * from "./post";
