+++ src/graphql/mutations/index.js
@@ -1,0 +1,2 @@
+export * from "./delete-movie";
+export * from "./update-movie-status";
